fixed an error

+++ intlist.cpp
@@ -75,9 +75,9 @@ int IntList::max() const {
 // 0 if list is empty
 double IntList::average() const {
     if (head==nullptr){
-        return 0.000;
+        return 0.0000;
     }
-    return sum()/count(); // REPLACE THIS NON-SOLUTION
+    return static_cast<double>(sum()/count()); // REPLACE THIS NON-SOLUTION
 }
 
 // inserts value as new node at beginning of list
